requirements.txt
langchain
chromadb==0.4.24
openai
unstructured
fastapi
uvicorn
python-docx
PyMuPDF
pandas
langchain
geopandas
shapely
folium
tqdm
langchain_community
llama-index-vector-stores-chroma==0.10.30
python-dotenv
python-multipart
numpy==1.26.4
transformers
sentence-transformers
accelerate
llama-index-embeddings-huggingface
sentencepiece
huggingface_hub
auto-gptq
torch
llama-index-llms-huggingface
llama-index-llms-huggingface-api
